Server/tf/requirements.txt
Flask==1.1.2
Flask-WTF==0.14.3
WTForms==2.3.3
flask-mongoengine==0.9.5
Flask-Cors==3.0.8
python-dotenv==0.13.0
mongoengine~=0.20.0
Flask-RESTful==0.3.8
pymongo==3.10.1
redis==3.5.2
rq==1.4.1
nltk==3.4.5
Flask-Cache==0.13.1
Shapely==1.7.1
tensorflow==2.4.1
pandas==1.2.3